faceblur-man/requirements.txt
gradio>=4.29.0
numpy
Pillow
